Use size_t for the argument of operator[] for link collections

+++ include/podio/detail/LinkCollectionImpl.h
@@ -85,11 +85,11 @@ public:
   }
 
   /// Returns the immutable object of given index
-  value_type operator[](unsigned int index) const {
+  value_type operator[](size_t index) const {
     return value_type(m_storage.entries[index]);
   }
   /// Returns the mutable object of given index
-  mutable_type operator[](unsigned int index) {
+  mutable_type operator[](size_t index) {
     return mutable_type(podio::utils::MaybeSharedPtr(m_storage.entries[index]));
   }
   /// Returns the immutable object of given index
